[iris] Move native-proxy metrics into their own module
The native listener's iris_rpc_* and iris_proxy_* metric families —
counters, latency histograms, per-request timers, and their tests — had
accreted inside lib.rs. Move them to a dedicated metrics module. The
timers now take the metric store they record into rather than the whole
ProxyControl, leaving the module with no dependency back on the listener.
Behavior and the Python-facing snapshot JSON are unchanged.
